requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
pydantic
google-generativeai
python-dotenv
pypdf
python-docx